app/requirements.txt
Django==1.11
psycopg2
djangorestframework==3.7.3
djangorestframework-jwt==1.11.0